FarmaciaDigital/requirements.txt
django==4.1
pillow
pymysql
mysqlclient
django_mysql
django-jsonfield
django-json-field
django-import-export
django-admin-interface
django-crispy-forms
python-dotenv
python-decouple
django-bootstrap4
django-environ
cx-oracle
django-rest-framework
requests
reportlab
googlemaps